Render case studies in government-frontend
We have been publishing case studies to publishing-api for some time, and
now that government-frontend is ready to handle case studies so we can finally
use it to render them in production.

Without the migration to republish all existing case studies with the new
rendering_app, newly published case studies would be served by government-frontend,
but existing ones would stay in Whitehall.